Add reaction responder, working removing of messages

+++ src/Accord.Services/UserBotMessages/DeleteUserBotMessage.cs
@@ -8,32 +8,32 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Accord.Services.UserBotMessages;
 
-public sealed record DeleteUserBotMessageRequest(PermissionUser User, ulong DiscordMessageId) : IRequest<ServiceResponse>;
+public sealed record DeleteUserBotMessageRequest(PermissionUser User, ulong DiscordMessageId) : INotification;
+
 public sealed record DeleteDiscordMessageRequest(ulong DiscordMessageId, ulong DiscordChannelId) : IRequest<ServiceResponse>;
 
-public class DeleteUserBotMessage(AccordContext db, IMediator mediator) : IRequestHandler<DeleteUserBotMessageRequest, ServiceResponse>
+public class DeleteUserBotMessage(AccordContext db, IMediator mediator) : INotificationHandler<DeleteUserBotMessageRequest>
 {
-    public async Task<ServiceResponse> Handle(DeleteUserBotMessageRequest request, CancellationToken cancellationToken)
+    public async Task Handle(DeleteUserBotMessageRequest request, CancellationToken cancellationToken)
     {
         var message = await db.UserBotMessages
             .Where(x => x.Id == request.DiscordMessageId)
             .SingleOrDefaultAsync(cancellationToken: cancellationToken);
 
         if (message is null)
-        {
-            return ServiceResponse.Fail("The message either does not exist or cannot be removed");
-        }
+            return;
 
         if (message.UserId != request.User.DiscordUserId && !request.User.IsAdministrator)
         {
-            return ServiceResponse.Fail("The message can only be removed by the original initiator or administrators");
+            return;
         }
 
-        await mediator.Send(new DeleteDiscordMessageRequest(message.Id, message.DiscordChannelId), cancellationToken);
+        var response = await mediator.Send(new DeleteDiscordMessageRequest(message.Id, message.DiscordChannelId), cancellationToken);
+
+        if (!response.Success)
+            return;
 
         db.Remove(message);
         await db.SaveChangesAsync(cancellationToken);
-
-        return ServiceResponse.Ok();
     }
 }
